option to generate pdf outline
This depends on a recent version of wkhtmltopdf-static, and the outline
can be lost through further manipulation, like page numbering, so it is
only used for web-destined pdfs.

NOTE: because the outline can be read with pdftk dump_data, it makes
a better way of finding contents than is currently used.