Add option to embed standard PDF fonts
The standard fonts (Times, Helvetica, Courier, and others you don't
want to know about) are built into PDF viewers and are not usually
required to be embedded in the document.  But printers do things
differently.  When they say PDF, they mean PDF/X-3 or something
similar, and they do like all fonts to be embedded.

These fonts are no longer specified in Objavi CSS, embedding is not
required for most uses, this takes time and potentially risks mucking
things up, so embedding is off by default.