Workaround filenames in Xinha local links (#345)
Xinha produces document local links (e.g., for footnotes) in the form
'filename#local_anchor', which are broken if the filename changes.  In
practice the filename changes at least twice during processing -- once
from 'filename' to 'filename.html', when Booki makes the bookizip, and
again to 'body.html' when all the chapters get concatenated.

Additionally, Xinha will reuse the same IDs in each chapter, so when
the chapters are all concatenated the IDs are no longer unique and the
links won't work properly.

This commit adds a function (the documentation string of which this
commit comment is lifted from), that replaces links in the form
'filename#id' with '#filename_id', and change the target IDs
according.  It avoids altering the ID of elements that aren't locally
linked, as these might be used for CSS or external links.